# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: TravelBudget API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /preferences:
    get:
      operationId: getPreferences
      tags: [preferences]
      summary: Get user preferences
      responses:
        "200":
          description: User preferences
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserPreferences"
        "404":
          description: No preferences set yet
    put:
      operationId: savePreferences
      tags: [preferences]
      summary: Save or update user preferences
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserPreferencesInput"
      responses:
        "200":
          description: Preferences saved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserPreferences"
...
  /subscription/upgrade:
    post:
      operationId: upgradeSubscription
      tags: [preferences]
      summary: Upgrade to premium plan
      responses:
        "200":
          description: Subscription upgraded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SubscriptionStatus"
        "401":
          description: Unauthorized
  /subscription/downgrade:
    post:
      operationId: downgradeSubscription
      tags: [preferences]
      summary: Downgrade to free plan
      responses:
        "200":
          description: Subscription downgraded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SubscriptionStatus"
        "401":
          description: Unauthorized
components:
  schemas:
...
    UserPreferencesInput:
      type: object
      properties:
        language:
          type: string
        defaultBudget:
          type: ["number", "null"]
        defaultNumberOfPeople:
          type: ["integer", "null"]
        defaultDepartureLocation:
          type: ["string", "null"]
        defaultFlightPreference:
          type: ["string", "null"]
        cookieConsent:
          type: boolean
        onboardingCompleted:
          type: boolean
    UserPreferences:
      type: object
      required:
        - id
        - clerkUserId
      properties:
        id:
          type: integer
        clerkUserId:
          type: string
        language:
          type: string
        defaultBudget:
          type: ["number", "null"]
        defaultNumberOfPeople:
          type: ["integer", "null"]
        defaultDepartureLocation:
          type: ["string", "null"]
        defaultFlightPreference:
          type: ["string", "null"]
        cookieConsent:
          type: boolean
        onboardingCompleted:
          type: boolean
        createdAt:
          type: string
...
    SubscriptionStatus:
      type: object
      required:
        - isPremium
        - plan
      properties:
        isPremium:
          type: boolean
        plan:
          type: string
          enum: [free, premium]
...