# lib/api-zod/src/generated/types/subscriptionStatus.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * TravelBudget API specification
 * OpenAPI spec version: 0.1.0
 */
import type { SubscriptionStatusPlan } from './subscriptionStatusPlan';

export interface SubscriptionStatus {
  isPremium: boolean;
  plan: SubscriptionStatusPlan;
}